+++ src/renderer/components/OfflineToast.tsx
@@ -4,6 +4,7 @@ import { onDCEvent } from '../ipc'
 
 import { getLogger } from '../../shared/logger'
 import { useTranslationFunction } from '../contexts'
+import { useKeyBindingAction, KeybindAction } from '../keybindings'
 
 const log = getLogger('renderer/components/OfflineToast')
 
@@ -69,11 +70,18 @@ export default function OfflineToast() {
     setNetworkState(() => [true, ''])
   }
 
+  const maybeNetwork = () => {
+    DeltaBackend.call('context.maybeNetwork')
+  }
+
+  useKeyBindingAction(KeybindAction.Debug_MaybeNetwork, maybeNetwork)
+
   useEffect(() => {
     if (navigator.onLine === false) onBrowserOffline()
 
     window.addEventListener('online', onBrowserOnline)
     window.addEventListener('offline', onBrowserOffline)
+    window.addEventListener('focus', maybeNetwork)
     const removeEventListenerDCNetworkError = onDCEvent(
       'DC_EVENT_ERROR_NETWORK',
       onDeltaNetworkError
@@ -86,6 +94,7 @@ export default function OfflineToast() {
     return () => {
       window.removeEventListener('online', onBrowserOffline)
       window.removeEventListener('offline', onBrowserOnline)
+      window.removeEventListener('focus', maybeNetwork)
 
       removeEventListenerDCNetworkError()
       removeEventListenerDCNetworkSuccess()